pyproject.toml
[tool.poetry]
name = "meli-ldap-agent"
version = "0.1.0"
description = "Sistema de Agentes LDAP Auto-Adaptativos para Mercado Libre"
authors = ["Tu Nombre"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
python-ldap = "^3.4.3"
langchain = "^0.3.0"
langchain-google-genai = "^2.0.0"
langgraph = "^0.2.0"
pydantic = "^2.5.0"
python-dotenv = "^1.0.0"
google-generativeai = "^0.8.0"
google-genai = "^1.72.0"
litellm = "^1.40.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
